Add support for `import.meta.url` & `import.meta.resolve()`
Also switches from using `rollup` and `eslint` as dev dependencies
to using `@shgysk8zer0/js-utils`.
