+++ .npmignore
@@ -1,5 +1,6 @@
 node_modules/
 test/
+.env
 .github/
 .editorconfig
 .eslintignore
@@ -14,4 +15,6 @@ rollup.config.js
 test.config.js
 *.tgz
 *.log
-*.out.js
+*.bak
+*.min.js
+*.map
