+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shgysk8zer0/rollup-import",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "description": "A RollUp plugin for importing modules from URLs, paths, and bare specifiers using import maps.",
   "type": "module",
   "engines": {
@@ -13,6 +13,9 @@
     ".": {
       "import": "./index.js",
       "require": "./index.cjs"
+    },
+    "./*": {
+      "import": "./*.js"
     }
   },
   "scripts": {
@@ -46,14 +49,15 @@
     "url": "https://github.com/shgysk8zer0/rollup-import/issues"
   },
   "homepage": "https://github.com/shgysk8zer0/rollup-import#readme",
+  "dependencies": {
+    "@shgysk8zer0/npm-utils": "^1.0.4",
+    "magic-string": "^0.30.0"
+  },
   "devDependencies": {
-    "eslint": "^8.40.0",
-    "rollup": "^3.2.2"
+    "@shgysk8zer0/js-utils": "^1.0.0",
+    "dotenv": "^16.1.4"
   },
   "peerDependencies": {
     "rollup": "*"
-  },
-  "dependencies": {
-    "@shgysk8zer0/npm-utils": "^1.0.4"
   }
 }
